+++ Sorting/time.h
@@ -13,7 +13,7 @@
  * \param to_low true if sort from high to low
  * \return time in seconds
  */
-std::chrono::duration<double, std::ratio<1, 1>> time(void (*f)(dualtype&, const uint64_t& col, const bool&), dualtype& a, const uint64_t& col, const bool& to_low);
+std::chrono::duration<double, std::ratio<1, 1>> time(void (*f)(dualtype&, const uint16_t& col, const bool&), dualtype& a, const uint16_t& col, const bool& to_low);
 /**
  * \brief this function testing time for function complete in [n] iterations
  * \param f pointer to function
@@ -24,5 +24,5 @@ std::chrono::duration<double, std::ratio<1, 1>> time(void (*f)(dualtype&, const 
  * \param to_low true if sort from high to low
  * \return time in seconds
  */
-std::chrono::duration<double, std::ratio<1, 1>> test(void (*f)(dualtype&, const uint64_t& col, const bool&), dualtype& a, const uint16_t& n, const address& path, const uint64_t& col, const bool& to_low);
-void full_test(dualtype& a, const uint16_t& n, const address& path, const uint64_t& col, const bool& to_low);
+std::chrono::duration<double, std::ratio<1, 1>> test(void (*f)(dualtype&, const uint16_t& col, const bool&), dualtype& a, const uint16_t& n, const address& path, const uint16_t& col, const bool& to_low);
+void full_test(dualtype& a, const uint16_t& n, const address& path, const uint16_t& col, const bool& to_low);
